Re-enabled encoders, added method for distance.

+++ lib/movement_lib/movement.h
@@ -25,4 +25,4 @@ void pravyMotorStop();
 void pohyb(int rychlostL, int rychlostR);
 bool otacej_dokud_nenajdes_caru(byte position, int8_t smer);
 void otacej_dle_offsetu(int offset);
-void turn(int angle, int8_t smer);
+void turn(byte angle, int8_t smer);
